+++ dtdl2-validator/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "dtdl2-validator": {
       "commandName": "Project",
-      "commandLineArgs": "..\\..\\..\\..\\test\\demodevice.json"
+      "commandLineArgs": "/f=..\\..\\..\\..\\test\\demodevice.json"
     }
   }
 }
